fix-修改分页获取权限接口的分页逻辑

+++ adaptor/repo/dto/page.go
@@ -1,8 +1,8 @@
 package dto
 
 type PageDto struct {
-	PageNum  int    `json:"page_num"`
-	PageSize int    `json:"page_size"`
+	PageNum  int64  `json:"page_num"`
+	PageSize int64  `json:"page_size"`
 	Name     string `json:"name"`
 	Sort     int    `json:"sort"`
 	CreateBy int64  `json:"create_by"`
